pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=3.0.1",
    "alembic>=1.13.2",
    "celery>=5.3.4",
    "email-validator>=2.1.0",
    "fastapi[standard]>=0.116.1",
    "jinja2>=3.1.2",
    "passlib[argon2]>=1.3",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "redis>=5.0.1",
    "sqlalchemy>=2.0.35",
]
